Enum r2d2_mysql::param::ConnectError

pub enum ConnectError {
    InvalidUrl(String),
    MissingUser,
    MissingPassword,
    UnsupportedAuthentication,
    NoSslSupport,
    SslError(Box<Error + Sync + Send>),
    IoError(Error),
}

Reasons a new Postgres connection could fail.

Variants

InvalidUrl

The provided URL could not be parsed.

MissingUser

The URL was missing a user.

MissingPassword

A password was required but not provided in the URL.

UnsupportedAuthentication

The Postgres server requested an authentication method not supported by the driver.

NoSslSupport

The Postgres server does not support SSL encryption.

SslError

There was an error initializing the SSL session.

IoError

There was an error communicating with the server.